requirements_generation.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cobble==0.1.4
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
datasets==3.0.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.30.2
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==1.2.0
face_detection==0.2.2
fastapi==0.114.1
filelock==3.16.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.34.0
google_search_results==2.4.2
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
hydra-core==1.3.2
idna==3.8
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
iopath==0.1.10
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.7
kubernetes==30.1.0
langchain==0.2.16
langchain-chroma==0.1.3
langchain-community==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.118
lazy_loader==0.4
lmdb==1.5.1
loguru==0.7.2
lxml==5.3.0
mammoth==1.8.0
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.44.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt-einsum==3.3.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle-gpu==2.6.1
pandas==2.2.2
parso==0.8.3
pathvalidate==3.2.1
pdfkit==1.0.0
pdfminer.six==20240706
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.2
portalocker==2.10.1
posthog==3.6.5
prompt-toolkit==3.0.39
protobuf 
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
puremagic==1.27
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.0
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
# SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git@7e1596c0b6462eb1d1ba7e1492430fed95023598
scikit-image==0.24.0
scipy==1.13.1
serpapi==0.1.5
shapely==2.0.6
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.2
starlette==0.38.5
sympy==1.13.2
tenacity==8.5.0
termcolor==2.4.0
tifffile==2024.8.30
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tornado==6.3.2
tqdm==4.66.5
traitlets==5.9.0
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.11.1
youtube-transcript-api==0.6.2
zipp==3.20.1
moviepy==1.0.3
gtts==2.5.3
soundfile==0.12.1
